+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/BlockstackSession.kt
@@ -8,35 +8,37 @@ import android.util.Log
 import android.webkit.JavascriptInterface
 import android.webkit.WebView
 import android.webkit.WebViewClient
-import org.json.JSONArray
 import org.json.JSONObject
-import java.net.URI
+import java.net.URL
 import java.util.*
 
 private val AUTH_URL_STRING = "file:///android_res/raw/webview.html"
 private val HOSTED_BROWSER_URL_BASE = "https://browser.blockstack.org"
 
-class BlockstackSession(private val context: Context,
-                        private val appDomain: URI,
-                        private val redirectURI: URI,
-                        private val manifestURI: URI,
-                        private val scopes: Array<Scope>,
+class BlockstackSession(context: Context,
+                        private val config: BlockstackConfig,
                         onLoadedCallback: () -> Unit = {}) {
 
     private val TAG = BlockstackSession::class.qualifiedName
     var loaded: Boolean = false
-        private set(value) {field = value}
+        private set(value) {
+            field = value
+        }
 
     private var userData: JSONObject? = null
     private var signInCallback: ((UserData) -> Unit)? = null
+    private val lookupProfileCallbacks = HashMap<String, ((Profile) -> Unit)>()
     private val getFileCallbacks = HashMap<String, ((Any) -> Unit)>()
     private val putFileCallbacks = HashMap<String, ((String) -> Unit)>()
+
 
 
     init {
         Log.d(TAG, context.toString())
     }
+
     private val webView = WebView(context)
+
     init {
         webView.settings.javaScriptEnabled = true
         webView.settings.domStorageEnabled = true
@@ -44,7 +46,7 @@ class BlockstackSession(private val context: Context,
             this.loaded = true
             onLoadedCallback()
         }
-        webView.addJavascriptInterface(JavascriptInterfaceObject(this),"android")
+        webView.addJavascriptInterface(JavascriptInterfaceObject(this), "android")
         webView.loadUrl(AUTH_URL_STRING)
     }
 
@@ -61,7 +63,7 @@ class BlockstackSession(private val context: Context,
         ensureLoaded()
 
         val javascript = "handlePendingSignIn('${authResponse}')"
-        webView.evaluateJavascript(javascript, { result: String ->
+        webView.evaluateJavascript(javascript, { _: String ->
 
         })
     }
@@ -74,15 +76,15 @@ class BlockstackSession(private val context: Context,
      * when authentication succeeds. It is not called on the UI thread so you should
      * execute any UI interactions in a `runOnUIThread` block
      */
-    fun redirectUserToSignIn(signInCallback: (UserData) -> Unit ) {
+    fun redirectUserToSignIn(signInCallback: (UserData) -> Unit) {
         this.signInCallback = signInCallback
         Log.d(TAG, "redirectUserToSignIn")
 
         ensureLoaded()
 
-        val scopesString = Scope.scopesArrayToJSONString(scopes)
-        val javascript = "redirectToSignIn('${appDomain}', '${redirectURI}', '${manifestURI}', ${scopesString})"
-        webView.evaluateJavascript(javascript, { result: String ->
+        val scopesString = Scope.scopesArrayToJSONString(config.scopes)
+        val javascript = "redirectToSignIn('${config.appDomain}', '${config.redirectURI}', '${config.manifestURI}', ${scopesString})"
+        webView.evaluateJavascript(javascript, { _: String ->
             // no op
         })
     }
@@ -98,14 +100,14 @@ class BlockstackSession(private val context: Context,
 
         ensureLoaded()
 
-        webView.evaluateJavascript(javascript, {result ->
-                if (result != null) {
-                    val newUserData = JSONObject(result)
-                    userData = newUserData
-                    callback(UserData(newUserData))
-                } else {
-                    callback(null)
-                }
+        webView.evaluateJavascript(javascript, { result ->
+            if (result != null && !"null".equals(result)) {
+                val newUserData = JSONObject(result)
+                userData = newUserData
+                callback(UserData(newUserData))
+            } else {
+                callback(null)
+            }
         })
     }
 
@@ -141,6 +143,25 @@ class BlockstackSession(private val context: Context,
         })
     }
 
+    fun lookupProfile(username: String, zoneFileLookupURL: URL? = null, callback: (Profile?) -> Unit) {
+        val javascript = if (zoneFileLookupURL != null) {
+            "lookupProfile('$username', '$zoneFileLookupURL')"
+        } else {
+            "lookupProfile('$username')"
+        }
+        ensureLoaded()
+        lookupProfileCallbacks.put(username, callback)
+        webView.evaluateJavascript(javascript, {
+            result ->
+            if (result != null && !"null".equals(result)) {
+                val newUserData = JSONObject(result)
+                callback(Profile(newUserData))
+            } else {
+                callback(null)
+            }
+        })
+    }
+
     /* Public storage methods */
 
     /**
@@ -159,7 +180,7 @@ class BlockstackSession(private val context: Context,
 
         val uniqueIdentifier = addGetFileCallback(callback)
         val javascript = "getFile('${path}', ${options}, '${uniqueIdentifier}')"
-        webView.evaluateJavascript(javascript, { result: String ->
+        webView.evaluateJavascript(javascript, { _: String ->
             // no op
         })
     }
@@ -181,26 +202,85 @@ class BlockstackSession(private val context: Context,
         ensureLoaded()
 
         val valid = content is String || content is ByteArray
-        if(!valid) {
+        if (!valid) {
             throw IllegalArgumentException("putFile content only supports String or ByteArray")
         }
 
         val isBinary = content is ByteArray
 
         val uniqueIdentifier = addPutFileCallback(callback)
-        if(isBinary) {
+        if (isBinary) {
             val contentString = Base64.encodeToString(content as ByteArray, Base64.NO_WRAP)
             val javascript = "putFile('${path}', '${contentString}', ${options}, '${uniqueIdentifier}', true)"
-            webView.evaluateJavascript(javascript, { result: String ->
+            webView.evaluateJavascript(javascript, { _: String ->
                 // no op
             })
         } else {
             val javascript = "putFile('${path}', '${content}', ${options}, '${uniqueIdentifier}', false)"
-            webView.evaluateJavascript(javascript, { result: String ->
+            webView.evaluateJavascript(javascript, { _: String ->
                 // no op
             })
         }
 
+    }
+
+    fun encryptContent(plainContent: Any, options: CryptoOptions, callback: (CipherObject?) -> Unit) {
+        ensureLoaded()
+
+        val valid = plainContent is String || plainContent is ByteArray
+        if (!valid) {
+            throw IllegalArgumentException("encrypt content only supports String or ByteArray")
+        }
+
+        val isBinary = plainContent is ByteArray
+
+        val javascript = if (isBinary) {
+            val contentString = Base64.encodeToString(plainContent as ByteArray, Base64.NO_WRAP)
+            "encryptContent('$contentString', $options, true)"
+        } else {
+            "encryptContent('$plainContent', $options, false)"
+        }
+
+        webView.evaluateJavascript(javascript) { result ->
+            if (result != null && !"null".equals(result)) {
+                val cipherObject = JSONObject(result)
+                callback(CipherObject(cipherObject))
+            } else {
+                callback(null)
+            }
+        }
+    }
+
+    fun decryptContent(cipherObject: Any, options: CryptoOptions, callback: (Any) -> Unit) {
+        ensureLoaded()
+
+        val valid = cipherObject is String || cipherObject is ByteArray
+        if (!valid) {
+            throw IllegalArgumentException("decrypt content only supports JSONObject or ByteArray")
+        }
+
+        val isBinary = cipherObject is ByteArray
+
+        var wasString:Boolean
+
+        val javascript = if (isBinary) {
+            val cipherTextString = Base64.encodeToString(cipherObject as ByteArray, Base64.NO_WRAP)
+            wasString = JSONObject(cipherTextString).getBoolean("wasString")
+            "decryptContent('$cipherTextString', $options, true)"
+        } else {
+            wasString = JSONObject(cipherObject as String).getBoolean("wasString")
+            "decryptContent('$cipherObject', $options, false)"
+        }
+
+
+        webView.evaluateJavascript(javascript) {plainContent ->
+
+            if (wasString) {
+                callback(plainContent)
+            } else {
+                callback(Base64.decode(plainContent, Base64.DEFAULT))
+            }
+        }
     }
 
     private fun addGetFileCallback(callback: (Any) -> Unit): String {
@@ -216,7 +296,7 @@ class BlockstackSession(private val context: Context,
     }
 
     private fun ensureLoaded() {
-        if(!this.loaded) {
+        if (!this.loaded) {
             throw IllegalStateException("Blockstack session hasn't finished loading." +
                     " Please wait until the onLoadedCallback() has fired before performing operations.")
         }
@@ -226,19 +306,25 @@ class BlockstackSession(private val context: Context,
 
         @JavascriptInterface
         fun signInSuccess(userDataString: String) {
-            Log.d(session.TAG, "signInSuccess" )
+            Log.d(session.TAG, "signInSuccess")
             val userData = JSONObject(userDataString)
             session.userData = userData
-            Log.d(session.TAG, session.userData.toString() )
+            Log.d(session.TAG, session.userData.toString())
             session.signInCallback?.invoke(UserData(userData))
         }
 
         @JavascriptInterface
+        fun lookupProfileResult(username:String, userDataString:String) {
+            val userData = JSONObject(userDataString)
+            session.lookupProfileCallbacks[username]?.invoke(Profile(userData))
+        }
+
+        @JavascriptInterface
         fun getFileResult(content: String, uniqueIdentifier: String, isBinary: Boolean) {
-            Log.d(session.TAG, "putFileResult" )
+            Log.d(session.TAG, "putFileResult")
 
             if (isBinary) {
-                val binaryContent: ByteArray = Base64.decode(content as String, Base64.DEFAULT)
+                val binaryContent: ByteArray = Base64.decode(content, Base64.DEFAULT)
                 session.getFileCallbacks[uniqueIdentifier]?.invoke(binaryContent)
             } else {
                 session.getFileCallbacks[uniqueIdentifier]?.invoke(content)
@@ -248,7 +334,7 @@ class BlockstackSession(private val context: Context,
 
         @JavascriptInterface
         fun putFileResult(readURL: String, uniqueIdentifier: String) {
-            Log.d(session.TAG, "putFileResult" )
+            Log.d(session.TAG, "putFileResult")
 
             session.putFileCallbacks[uniqueIdentifier]?.invoke(readURL)
             session.putFileCallbacks.remove(uniqueIdentifier)
@@ -259,7 +345,7 @@ class BlockstackSession(private val context: Context,
 }
 
 private class BlockstackWebViewClient(val context: Context,
-                                      val onLoadedCallback: () -> Unit ) : WebViewClient() {
+                                      val onLoadedCallback: () -> Unit) : WebViewClient() {
     private val TAG = BlockstackWebViewClient::class.qualifiedName
 
     override fun onPageFinished(view: WebView?, url: String?) {