+++ example-multi-activities/src/main/java/org/blockstack/android/AccountActivity.kt
@@ -10,6 +10,7 @@ import android.view.MenuItem
 import android.view.View
 import kotlinx.android.synthetic.main.activity_account.*
 import kotlinx.android.synthetic.main.content_account.*
+import org.blockstack.android.sdk.BlockstackConfig
 import org.blockstack.android.sdk.BlockstackSession
 import org.blockstack.android.sdk.Scope
 import java.net.URI
@@ -29,17 +30,13 @@ class AccountActivity : AppCompatActivity() {
         signInButton.isEnabled = false
         signOutButton.isEnabled = false
 
-        val appDomain = URI("https://flamboyant-darwin-d11c17.netlify.com")
-        val redirectURI = URI("${appDomain}/redirect")
-        val manifestURI = URI("${appDomain}/manifest.json")
-        val scopes = arrayOf(Scope.StoreWrite)
-
-        _blockstackSession = BlockstackSession(this, appDomain, redirectURI, manifestURI, scopes,
+        _blockstackSession = BlockstackSession(this, defaultConfig,
                 onLoadedCallback = {
                     if (intent?.action == Intent.ACTION_VIEW) {
                         handleAuthResponse(intent)
                     }
-                    onLoaded()})
+                    onLoaded()
+                })
 
         signInButton.setOnClickListener { view: View ->
             blockstackSession().redirectUserToSignIn { userData ->
@@ -113,13 +110,14 @@ class AccountActivity : AppCompatActivity() {
         return super.onOptionsItemSelected(item)
     }
 
-    fun blockstackSession() : BlockstackSession {
+    fun blockstackSession(): BlockstackSession {
         val session = _blockstackSession
-        if(session != null) {
+        if (session != null) {
             return session
         } else {
             throw IllegalStateException("No session.")
         }
     }
-
 }
+
+