+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - A `BlockstackConfig` object to make it easier to use across several activities. Thanks
 to @friedger.
+- `BlockstackSession.lookupProfile` method
 ### Changed
 - Fixed a bug where loadUserData would throw an exception if the user is not logged in.
 Thanks to @friedger.